refactor(#1052): SonarCloud pass — split the complexity hotspots, fix the blocker
Behaviour-neutral (99 tests unchanged):
- BLOCKER S912: the `added.isEmpty() || ++m_plannerRetries > ...` side effect
  is a plain increment now (expandCapabilities + handlePlanReply)
- onPlannerCompleted (CC 48) → dispatcher + handlePlanReply / handleReplanReply /
  appendRepairedTail / expandCapabilities
- validateArguments (CC 105) → per-type coercers + matchEnum
- destructiveReason (CC 33) → deleteReason + overwriteReason
- observationFromToolResult (CC 53) → resultText / parseErrorStatus /
  parseLabelledNumbers / parseJsonFacts / parseArtifacts / parseWarnings
- summarize (CC 41) → summaryHeadline / successObservationFor / factsFragment
- extractJsonObject's lambda → firstBalancedObject free function
- C-style tables → std::vector; qsizetype narrowing casts; one-statement-
  per-line in the planner backend lambdas; nested ternaries unrolled;
  a header comment Sonar read as commented-out code reworded

+++ src/AICapabilityRegistry.h
@@ -43,7 +43,7 @@ public:
     explicit AICapabilityRegistry(const QJsonArray& toolList);
 
     bool isEmpty() const { return m_tools.isEmpty(); }
-    int  toolCount() const { return m_tools.size(); }
+    int  toolCount() const { return static_cast<int>(m_tools.size()); }
 
     QStringList capabilityIds() const;
     const Capability* capability(const QString& id) const;
